Refactored image code, percentage of control now displayed
`image.js` contains the code for rendering an image with `voronoi.js`
providing the shader and buffer information.

Percentage of the pitch controlled by the teams is now displayed when
both teams are active.  This is calculated using mipmap to reduce the
pitch image to a single pixel.  Hopefully this is reasonably accurate!